Finished implementation of layline action. Fixes #5.
Fixed layline on left side of target, and removed timer implementation of layline tacking.
Changed subscribers in navigator_node.py to use a queue size of 1 so after an action returns, we only use the most recent boat position.
Also got actions to return updated target_headings, and set target_heading to be cur_boat_heading in the case where we have to tack past our original desired heading.
Updated target_headings ensure that when the navigator starts planning after an action, it has up to date information to plan with.